backend/requirements.txt
annotated-types==0.7.0

anyio==4.6.2.post1
certifi==2024.8.30

charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
elevenlabs==1.12.1
fastapi==0.115.5
h11==0.14.0
httptools==0.5.0
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
pydantic==2.9.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.1